+++ components/TodoModal.tsx
@@ -1,78 +1,25 @@
-import axios from '@/lib/axios';
-import formatDate from '@/utils/formatDate';
+import { getCard } from '@/pages/api/getCard';
+import { Card } from '@/types/Card';
 import Image from 'next/image';
-import { ChangeEvent, SyntheticEvent, useEffect, useState } from 'react';
+import { useEffect, useState } from 'react';
 
+import Comments from './Comments';
 import Modal from './Modal';
 
-interface CardServiceResponseDto {
-  id: number;
-  title: string;
-  description: string;
-  tags: string[];
-  dueDate: string;
-  assignee: {
-    profileImageUrl: string;
-    nickname: string;
-    id: number;
-  };
-  imageUrl: string;
-  teamId: string;
-  columnId: number;
-  createdAt: string;
-  updatedAt: string;
-}
-
-interface Comment {
-  id: number;
-  content: string;
-  createdAt: string;
-  updatedAt: string;
-  cardId: number;
-  author: {
-    profileImageUrl: string;
-    nickname: string;
-    id: number;
-  };
-}
-
-interface CommentList {
-  cursorId: number;
-  comments: Comment[];
-}
-
 const TodoModal = () => {
-  const [card, setCard] = useState<CardServiceResponseDto>();
-  const [commentList, setCommentList] = useState<CommentList>();
+  const [card, setCard] = useState<Card>();
+
   const [kebabButtonVisible, setKebabButtonVisible] = useState(false);
   const [isModalOpen, setIsModalOpen] = useState(false);
-  const [inputValue, setInputValue] = useState('');
 
   const openModal = () => setIsModalOpen(true);
   const closeModal = () => setIsModalOpen(false);
 
   useEffect(() => {
     const getTodoCard = async () => {
-      const token =
-        'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6Mzk3NiwidGVhbUlkIjoiNi0zIiwiaWF0IjoxNzE5MjAwODQzLCJpc3MiOiJzcC10YXNraWZ5In0.ybVvT21thF6vjcG5ReI_XlIHCSn45HoFt6FTWKYYAm8';
-      try {
-        const res = await axios.get('cards/8667', {
-          headers: {
-            Authorization: `Bearer ${token}`,
-          },
-        });
-        const commentsRes = await axios.get('comments?size=1&cardId=8667', {
-          headers: {
-            Authorization: `Bearer ${token}`,
-          },
-        });
-        const data: CardServiceResponseDto = res.data;
-        const commentsData: CommentList = commentsRes.data;
-        setCard(data);
-        setCommentList(commentsData);
-      } catch (error) {
-        console.error('Failed to fetch data:', error);
-      }
+      const cardData: Card = await getCard();
+
+      setCard(cardData);
     };
 
     getTodoCard();
@@ -85,14 +32,6 @@ const TodoModal = () => {
   const handleEditButton = () => {};
 
   const handleDeleteButton = () => {};
-
-  const handleInputChange = (e: ChangeEvent<HTMLTextAreaElement>) => {
-    setInputValue(e.target.value);
-  };
-
-  const handleSubmit = (e: SyntheticEvent<HTMLFormElement>) => {
-    e.preventDefault();
-  };
 
   return (
     <div className="px-28 py-32">
@@ -132,58 +71,8 @@ const TodoModal = () => {
                   fill
                 />
               </div>
-              <form
-                onSubmit={handleSubmit}
-                className="relative flex flex-col gap-10"
-              >
-                <div className="text-16 font-medium leading-[19px] text-black-20">
-                  댓글
-                </div>
-                <div>
-                  <textarea
-                    value={inputValue}
-                    onChange={handleInputChange}
-                    placeholder="댓글 작성하기"
-                    className="w-450 h-110 resize-none rounded-6 border border-solid border-gray-30 p-16 placeholder:text-14"
-                  />
-                </div>
-                <button className="btn_desktop_white absolute bottom-16 left-[353px]">
-                  입력
-                </button>
-              </form>
 
-              {commentList &&
-                commentList.comments.map((comment) => (
-                  <div className="flex flex-col" key={comment.id}>
-                    <div className="mt-20 flex gap-10">
-                      <div className="flex justify-start">
-                        <Image
-                          width={34}
-                          height={34}
-                          src={'/images/icon/younghoon.svg'}
-                          alt="담당자 프로필 이미지"
-                        />
-                      </div>
-                      <div className="flex items-center gap-8">
-                        <div className="text-14 font-semibold leading-[17px] text-black-20">
-                          {comment.author.nickname}
-                        </div>
-                        <div className="text-12 font-normal leading-[14px] text-gray-40">
-                          {formatDate(comment.createdAt)}
-                        </div>
-                      </div>
-                    </div>
-                    <div className="ml-44 flex flex-col gap-12">
-                      <div className="text-14 font-normal leading-[17px] text-black-20">
-                        {comment.content}
-                      </div>
-                      <div className="flex gap-12 text-12 leading-[14px] text-gray-40 underline">
-                        <div>수정</div>
-                        <div>삭제</div>
-                      </div>
-                    </div>
-                  </div>
-                ))}
+              <Comments />
             </div>
             <div className="flex flex-col gap-21">
               <div className="relative flex justify-end gap-24">
