+++ types/comments.ts
@@ -1,0 +1,17 @@
+export interface Comment {
+  id: number;
+  content: string;
+  createdAt: string;
+  updatedAt: string;
+  cardId: number;
+  author: {
+    profileImageUrl: string;
+    nickname: string;
+    id: number;
+  };
+}
+
+export interface CommentList {
+  cursorId: number;
+  comments: Comment[];
+}
