--- types/dummy.txt
@@ -1,1 +1,0 @@
-3팀
